requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Limiter==3.8.0
Werkzeug==3.0.3
itsdangerous==2.2.0
Gunicorn==22.0.0